% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_df3}
\alias{sim_df3}
\title{Simulated data with cumulative effects}
\format{An object of class \code{nested_fdf} (inherits from \code{sim_sdf}, \code{ped}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 250 rows and 9 columns.}
\usage{
sim_df3
}
\description{
This is data simulated using the \code{\link[pammtools]{sim_pexp}} function.
It contains two time-constant and two time-dependent covariates (observed
on different exposure time grids). The code used for simulation is
contained in the examples of \code{\link[pammtools]{sim_pexp}}.
}
\keyword{datasets}
